build: make clang-tidy and the compiler pass with zero findings
Passing rules printed nothing, so every compiler warning and every
clang-tidy finding in a non-strict check was thrown away. A rule whose
output contains a warning now prints it, and -Werror is on.

That surfaced 216 clang-tidy findings and a set of hidden test build
warnings. Fixed rather than suppressed where the code was wrong:

- nya_string_remove and nya_string_replace were declared with a non-const
  overload nothing defined.
- _nya_log_day_from_name parsed with sscanf, which accepts partial input.
- two reflection switches returned silently for NYA_REFLECT_COUNT.
- the curl, integrity, filesystem and render_sort headers formed include
  cycles.

Vendor includes are passed to clang-tidy as -isystem so third party code
is not reported. Macros taking type or declarator parameters carry a
NOLINT block, since those cannot be parenthesized, and the opt-in padding
check is off because the structs it flags are singletons or per-call
descriptors.

+++ src/nyangine/core/core_audio.c
@@ -1259,7 +1259,7 @@ void _nya_audio_reverb_apply(NYA_AudioReverbState* reverb, const SDL_AudioSpec* 
     const f32 allpass_feedback = 0.5F;
 
     for (s32 frame = 0; frame < frames; frame++) {
-        f32* row = &pcm[frame * channels];
+        f32* row = &pcm[(ptrdiff_t)frame * channels];
 
         /*
          * Downmixed into a left and a right feed by channel parity: even channels are the left side of
